@@ -1,8 +1,8 @@
 part of platform_maps_flutter;
 
 class PlatformMapController {
-  appleMaps.AppleMapController appleController;
-  googleMaps.GoogleMapController googleController;
+   appleMaps.AppleMapController? appleController;
+   googleMaps.GoogleMapController? googleController;
 
   PlatformMapController(dynamic controller) {
     if (controller.runtimeType == googleMaps.GoogleMapController) {
@@ -21,11 +21,10 @@ class PlatformMapController {
   ///   * [hideMarkerInfoWindow] to hide the Info Window.
   ///   * [isMarkerInfoWindowShown] to check if the Info Window is showing.
   Future<void> showMarkerInfoWindow(MarkerId markerId) {
-    assert(markerId != null);
     if (Platform.isAndroid) {
-      return googleController.showMarkerInfoWindow(markerId.googleMapsMarkerId);
+      return googleController!.showMarkerInfoWindow(markerId.googleMapsMarkerId);
     } else if (Platform.isIOS) {
-      return appleController
+      return appleController!
           .showMarkerInfoWindow(markerId.appleMapsAnnoationId);
     }
     throw ('Platform not supported.');
@@ -40,11 +39,10 @@ class PlatformMapController {
   ///   * [showMarkerInfoWindow] to show the Info Window.
   ///   * [isMarkerInfoWindowShown] to check if the Info Window is showing.
   Future<void> hideMarkerInfoWindow(MarkerId markerId) {
-    assert(markerId != null);
     if (Platform.isAndroid) {
-      return googleController.hideMarkerInfoWindow(markerId.googleMapsMarkerId);
+      return googleController!.hideMarkerInfoWindow(markerId.googleMapsMarkerId);
     } else if (Platform.isIOS) {
-      return appleController
+      return appleController!
           .hideMarkerInfoWindow(markerId.appleMapsAnnoationId);
     }
     throw ('Platform not supported.');
@@ -59,12 +57,11 @@ class PlatformMapController {
   ///   * [showMarkerInfoWindow] to show the Info Window.
   ///   * [hideMarkerInfoWindow] to hide the Info Window.
   Future<bool> isMarkerInfoWindowShown(MarkerId markerId) {
-    assert(markerId != null);
     if (Platform.isAndroid) {
-      return googleController
+      return googleController!
           .isMarkerInfoWindowShown(markerId.googleMapsMarkerId);
     } else if (Platform.isIOS) {
-      return appleController
+      return appleController!
           .isMarkerInfoWindowShown(markerId.appleMapsAnnoationId);
     }
     throw ('Platform not supported.');
@@ -76,9 +73,9 @@ class PlatformMapController {
   /// platform side.
   Future<void> animateCamera(cameraUpdate) async {
     if (Platform.isIOS) {
-      return this.appleController.animateCamera(cameraUpdate);
+      return this.appleController!.animateCamera(cameraUpdate);
     } else if (Platform.isAndroid) {
-      return this.googleController.animateCamera(cameraUpdate);
+      return this.googleController!.animateCamera(cameraUpdate);
     }
     throw ('Platform not supported.');
   }
@@ -89,22 +86,22 @@ class PlatformMapController {
   /// platform side.
   Future<void> moveCamera(cameraUpdate) async {
     if (Platform.isIOS) {
-      return this.appleController.moveCamera(cameraUpdate);
+      return this.appleController!.moveCamera(cameraUpdate);
     } else if (Platform.isAndroid) {
-      return this.googleController.moveCamera(cameraUpdate);
+      return this.googleController!.moveCamera(cameraUpdate);
     }
   }
 
   /// Return [LatLngBounds] defining the region that is visible in a map.
   Future<LatLngBounds> getVisibleRegion() async {
-    LatLngBounds _bounds;
+    late LatLngBounds _bounds;
     if (Platform.isIOS) {
       appleMaps.LatLngBounds appleBounds =
-          await this.appleController.getVisibleRegion();
+          await this.appleController!.getVisibleRegion();
       _bounds = LatLngBounds._fromAppleLatLngBounds(appleBounds);
     } else if (Platform.isAndroid) {
       googleMaps.LatLngBounds googleBounds =
-          await this.googleController.getVisibleRegion();
+          await this.googleController!.getVisibleRegion();
       _bounds = LatLngBounds._fromGoogleLatLngBounds(googleBounds);
     }
     return _bounds;
