@@ -10,7 +10,7 @@ part of platform_maps_flutter;
 @immutable
 class PolygonId {
   /// Creates an immutable identifier for a [Polygon].
-  PolygonId(this.value) : assert(value != null);
+  PolygonId(this.value);
 
   /// value of the [PolygonId].
   final String value;
@@ -18,7 +18,7 @@ class PolygonId {
   @override
   bool operator ==(Object other) {
     if (identical(this, other)) return true;
-    if (other.runtimeType != runtimeType) return false;
+    if (other is! PolygonId) return false;
     final PolygonId typedOther = other;
     return value == typedOther.value;
   }
@@ -37,7 +37,7 @@ class PolygonId {
 class Polygon {
   /// Creates an immutable representation of a polygon through geographical locations on the map.
   const Polygon({
-    @required this.polygonId,
+    required this.polygonId,
     this.consumeTapEvents = false,
     this.fillColor = Colors.black,
     this.points = const <LatLng>[],
@@ -77,18 +77,18 @@ class Polygon {
   final int strokeWidth;
 
   /// Callbacks to receive tap events for polygon placed on this map.
-  final VoidCallback onTap;
+  final VoidCallback? onTap;
 
   /// Creates a new [Polygon] object whose values are the same as this instance,
   /// unless overwritten by the specified parameters.
   Polygon copyWith({
-    bool consumeTapEventsParam,
-    List<LatLng> pointsParam,
-    Color strokeColorParam,
-    Color fillColorParam,
-    int strokeWidthParam,
-    bool visibleParam,
-    VoidCallback onTapParam,
+    bool? consumeTapEventsParam,
+    List<LatLng>? pointsParam,
+    Color? strokeColorParam,
+    Color? fillColorParam,
+    int? strokeWidthParam,
+    bool? visibleParam,
+    VoidCallback? onTapParam,
   }) {
     return Polygon(
       polygonId: polygonId,
@@ -110,7 +110,7 @@ class Polygon {
   @override
   bool operator ==(Object other) {
     if (identical(this, other)) return true;
-    if (other.runtimeType != runtimeType) return false;
+    if (other is! Polygon) return false;
     final Polygon typedOther = other;
     return polygonId == typedOther.polygonId &&
         consumeTapEvents == typedOther.consumeTapEvents &&
@@ -126,7 +126,7 @@ class Polygon {
   int get hashCode => polygonId.hashCode;
 
   static Set<googleMaps.Polygon> toGoogleMapsPolygonSet(Set<Polygon> polygons) {
-    List<googleMaps.Polygon> _polygons = [];
+    List<googleMaps.Polygon> _polygons = <googleMaps.Polygon>[];
     for (Polygon polygon in polygons) {
       _polygons.add(polygon.googleMapsPolygon);
     }
@@ -134,7 +134,7 @@ class Polygon {
   }
 
   static Set<appleMaps.Polygon> toAppleMapsPolygonSet(Set<Polygon> polygons) {
-    List<appleMaps.Polygon> _polygons = [];
+    List<appleMaps.Polygon> _polygons = <appleMaps.Polygon>[];
     for (Polygon polygon in polygons) {
       _polygons.add(polygon.appleMapsPolygon);
     }
